入力値制限を追加

+++ SmartAttendance/src/main/java/com/example/demo/form/RegistKintaiForm.java
@@ -1,6 +1,5 @@
 package com.example.demo.form;
 
-import java.math.BigDecimal;
 import java.sql.Date;
 
 import lombok.Data;
@@ -13,7 +12,7 @@ public class RegistKintaiForm {
 		private int start_M;
 		private int end_H;
 		private int end_M;
-		private BigDecimal total_workingTime;
+		private double total_workingTime;
 		private int breakTime; /*休憩時間[時間]*/
 		private int excessTime; /*超過労働時間[時間]*/
 }
